requirements.txt
Django==1.7.1
Pillow==2.8.1
reportlab==3.1.44
selenium==2.45.0
